+++ src/routes/prefs/+page.svelte
@@ -63,6 +63,18 @@
 			<b>Warning:</b> Disabling this will delete the existing cache.
 		</TogglePref>
 
+		<TogglePref
+			label="Fetch mods automatically"
+			value={prefs.fetchModsAutomatically}
+			set={set((value, prefs) => (prefs.fetchModsAutomatically = value))}
+		>
+			Whether to automatically start fetching mods when a game is selected and every 15 minutes
+			thereafter. This will ensure the mod list is up-to-date, but can be disabled to save
+			bandwidth.
+			<br />
+			To manually trigger a fetch, go to <b>File > Fetch mods</b>.
+		</TogglePref>
+
 		<Separator.Root class="h-2" />
 
 		<PathPref